refactor(discovery): modularize nexium dashboard & standardize UI
This commit overhauls the Nexium dashboard architecture by breaking down the monolithic 1800+ loc component into a clean, modular folder structure. Additionally, it addresses several UI inconsistencies, unifies brand iconography, and patches aggressive client-side caching bugs.

Core Architecture Changes:
- Extracted and modularized 'nexium-dashboard.tsx' into 'packages/components/profile/nexium/'
- Architected feature-specific components into a dedicated 'panels/' sub-directory (Signals, Opportunities, Skills, Applications, GitHub Repo Picker)
- Centralized Nexium constants and TypeScript types for seamless propagation across modules
- Patched TypeScript spread errors across the Nexium orchestrator by strictly narrowing loading states before applying profile spread updates

UI & Theming Standardization:
- Replaced all legacy native '<select>' tags with styled shadcn '<Select>' components across the Discovery dashboard and Squads roles interface
- Purged raw un-themed SVGs and completely standardized Discord branding by adopting '<SiDiscord>' in Linked Accounts, Profile Accounts, and Legal Hubs (replacing the generic MessageCircle icon)
- Fixed a layout shift on the Pricing page where expanding a tier card would uniformly stretch adjacent sibling cards
- Improved visual padding aesthetics within the Leaderboard UI

Performance & Synchronization Fixes:
- Patched a persistent desync bug where newly created API keys would fail to render instantly
- Injected dynamic timestamp cache-busters ('?t=Date.now()') across Personal API, Squad API, Domain, and Quota fetch routines to actively bypass strict Next.js router caches
