+++ packages/components/profile/accounts/linked-accounts.tsx
@@ -2,6 +2,7 @@
 
 import { useEffect, useState } from 'react'
 import { Github, Trash2, Link as LinkIcon, ImageIcon } from 'lucide-react'
+import { SiDiscord } from 'react-icons/si'
 import { Button } from '@/packages/components/ui/button'
 import { useToast } from '@/packages/hooks/use-toast'
 import { useRouter } from 'next/navigation'
@@ -210,13 +211,7 @@ export function LinkedAccounts() {
         <div className="flex flex-col gap-4 lg:flex-row lg:items-start lg:justify-between">
           <div className="flex items-start gap-3">
             <div className="p-2 rounded-lg bg-[#5865F2]/10">
-              <svg
-                className="h-5 w-5 text-[#5865F2]"
-                viewBox="0 0 24 24"
-                fill="currentColor"
-              >
-                <path d="M20.317 4.37a19.791 19.791 0 0 0-4.885-1.515a.074.074 0 0 0-.079.037c-.211.375-.444.864-.607 1.25a18.27 18.27 0 0 0-5.487 0c-.163-.386-.395-.875-.607-1.25a.077.077 0 0 0-.079-.037A19.736 19.736 0 0 0 3.677 4.37a.07.07 0 0 0-.032.027C.533 9.046-.32 13.58.099 18.057a.082.082 0 0 0 .031.057a19.9 19.9 0 0 0 5.993 3.03a.078.078 0 0 0 .084-.028a14.975 14.975 0 0 0 1.293-2.1a.07.07 0 0 0-.038-.098a13.11 13.11 0 0 1-1.872-.892a.072.072 0 0 1-.007-.12a10.149 10.149 0 0 0 .372-.294a.074.074 0 0 1 .076-.01c3.928 1.793 8.18 1.793 12.062 0a.074.074 0 0 1 .076.01c.12.098.246.198.373.294a.072.072 0 0 1-.006.12a12.296 12.296 0 0 1-1.873.892a.077.077 0 0 0-.037.099c.36.687.772 1.341 1.293 2.1a.076.076 0 0 0 .084.028a19.839 19.839 0 0 0 6.002-3.03a.077.077 0 0 0 .032-.054c.5-4.467.151-8.343-.434-12.033a.05.05 0 0 0-.02-.028zM8.02 15.33c-1.183 0-2.157-.965-2.157-2.156c0-1.193.974-2.157 2.157-2.157c1.193 0 2.156.964 2.157 2.157c0 1.19-.964 2.156-2.157 2.156zm7.975 0c-1.183 0-2.157-.965-2.157-2.156c0-1.193.974-2.157 2.157-2.157c1.193 0 2.156.964 2.157 2.157c0 1.19-.964 2.156-2.157 2.156z" />
-              </svg>
+              <SiDiscord className="h-5 w-5 text-[#5865F2]" />
             </div>
             <div className="space-y-1 flex-1">
               <h3 className="font-medium">Discord</h3>
