+++ packages/components/profile/public-profile.tsx
@@ -1,4 +1,4 @@
-﻿'use client'
+'use client'
 
 import { useState, useEffect } from 'react'
 import { Badge } from '@/packages/components/ui/badge'
